split quarkus extension test module to check configuration from both kmodule.xml and application.properties
