+++ drools-quarkus-extension/drools-quarkus-integration-test-kmodule/src/main/java/org/drools/quarkus/test/kmodule/ProbeCounter.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.drools.quarkus.test;
+package org.drools.quarkus.test.kmodule;
 
 public class ProbeCounter {
     private int total = 0;
