+++ drools-quarkus-extension/drools-quarkus-integration-test-kmodule/src/main/resources/META-INF/kmodule.xml
@@ -1,0 +1,31 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+    Licensed to the Apache Software Foundation (ASF) under one
+    or more contributor license agreements.  See the NOTICE file
+    distributed with this work for additional information
+    regarding copyright ownership.  The ASF licenses this file
+    to you under the Apache License, Version 2.0 (the
+    "License"); you may not use this file except in compliance
+    with the License.  You may obtain a copy of the License at
+      http://www.apache.org/licenses/LICENSE-2.0
+    Unless required by applicable law or agreed to in writing,
+    software distributed under the License is distributed on an
+    "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+    KIND, either express or implied.  See the License for the
+    specific language governing permissions and limitations
+    under the License.
+-->
+<kmodule xmlns="http://www.drools.org/xsd/kmodule" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+  <kbase name="cepKB" packages="org.drools.cep" eventProcessingMode="stream">
+    <ksession name="cepKS" clockType="pseudo"/>
+  </kbase>
+  <kbase name="tmsKB" packages="org.drools.tms">
+    <ksession name="tmsKS"/>
+  </kbase>
+  <kbase name="probeKB" packages="org.drools.probe" eventProcessingMode="stream">
+    <ksession name="probeKS"/>
+  </kbase>
+  <kbase name="allKB" packages="*" eventProcessingMode="stream">
+    <ksession name="allKS"/>
+  </kbase>
+</kmodule>
