+++ drools-quarkus-extension/drools-quarkus-integration-test-kmodule/src/main/resources/application.properties
@@ -17,17 +17,5 @@
 # under the License.
 #
 
-drools.prototypes=allowed
+quarkus.http.port=8080
 
-drools.kbase.canDrinkKB.packages=org.drools.drl
-drools.kbase.canDrinkKB.ksession.canDrinkKS.default=true
-
-drools.kbase.canDrinkKBDTable.packages=org.drools.dtable
-drools.kbase.canDrinkKBDTable.ksession=canDrinkKSDTable
-
-drools.kbase.canDrinkKBYaml.packages=org.drools.yaml
-drools.kbase.canDrinkKBYaml.ksession=canDrinkKSYaml
-
-drools.kbase.canDrinkKBPrototype.packages=org.drools.prototype
-drools.kbase.canDrinkKBPrototype.prototypes=allowed
-drools.kbase.canDrinkKBPrototype.ksession=canDrinkKSPrototype
