+++ drools-quarkus-extension/drools-quarkus-integration-test-kmodule/src/main/resources/org/drools/probe/probe.drl
@@ -1,7 +1,7 @@
 package org.drools.probe
 
-import org.drools.quarkus.test.ProbeEvent;
-import org.drools.quarkus.test.ProbeCounter;
+import org.drools.quarkus.test.kmodule.ProbeEvent;
+import org.drools.quarkus.test.kmodule.ProbeCounter;
 
 declare ProbeCounter
     @role (event)
