+++ drools-quarkus-extension/drools-quarkus-integration-test-kmodule/src/main/resources/org/drools/tms/tms.drl
