+++ drools-quarkus-extension/drools-quarkus-integration-test-kmodule/src/test/java/org/drools/quarkus/test/TestableIT.java
