+++ drools-quarkus-extension/pom.xml
@@ -40,6 +40,7 @@
     <module>drools-quarkus</module>
     <module>drools-quarkus-deployment</module>
     <module>drools-quarkus-integration-test</module>
+    <module>drools-quarkus-integration-test-kmodule</module>
     <module>drools-quarkus-integration-test-hotreload</module>
     <module>drools-quarkus-ruleunit-integration-test</module>
     <module>drools-quarkus-quickstart-test</module>
